refactor: 도메인별 패키지 분리
매핑 패키지만 두고 욱여넣었던 것들 각 역할별로 분산시켜서 패키지별로 저장
