+++ src/main/java/com/ctfplatform/backend/domain/challenge/Challenge.java
@@ -1,12 +1,8 @@
-package com.ctfplatform.backend.domain;
+package com.ctfplatform.backend.domain.challenge;
 
 import com.ctfplatform.backend.domain.common.BaseEntity;
 import com.ctfplatform.backend.domain.enums.Category;
 import com.ctfplatform.backend.domain.enums.Difficulty;
-import com.ctfplatform.backend.domain.mapping.ChallengeAuthor;
-import com.ctfplatform.backend.domain.mapping.ChallengeSubmission;
-import com.ctfplatform.backend.domain.mapping.SolveLog;
-import com.ctfplatform.backend.domain.mapping.TryChance;
 import jakarta.persistence.*;
 import lombok.*;
