+++ src/main/java/com/ctfplatform/backend/domain/challenge/ChallengeHint.java
@@ -1,4 +1,4 @@
-package com.ctfplatform.backend.domain;
+package com.ctfplatform.backend.domain.challenge;
 
 import com.ctfplatform.backend.domain.common.BaseEntity;
 import jakarta.persistence.*;
@@ -9,14 +9,19 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class ChallengeServer extends BaseEntity {
+public class ChallengeHint extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 100, updatable = false)
-    private String url;
+    @Lob
+    @Column(nullable = false)
+    private String hint;
+
+    @Builder.Default
+    @Column(nullable = false)
+    private Integer points = 0;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "challenge")
