+++ src/main/java/com/ctfplatform/backend/domain/challenge/ChallengeSubmission.java
@@ -1,24 +1,27 @@
-package com.ctfplatform.backend.domain.mapping;
+package com.ctfplatform.backend.domain.challenge;
 
-import com.ctfplatform.backend.domain.Challenge;
-import com.ctfplatform.backend.domain.Team;
-import com.ctfplatform.backend.domain.User;
+import com.ctfplatform.backend.domain.team.Team;
+import com.ctfplatform.backend.domain.user.User;
 import com.ctfplatform.backend.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class SolveLog extends BaseEntity {
+public class ChallengeSubmission extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(nullable = false, length = 100)
+    private String flag_hash;
+
+    @Column(nullable = false)
+    private Boolean correct;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user")
