+++ src/main/java/com/ctfplatform/backend/domain/challenge/TryChance.java
@@ -1,5 +1,6 @@
-package com.ctfplatform.backend.domain;
+package com.ctfplatform.backend.domain.challenge;
 
+import com.ctfplatform.backend.domain.team.Team;
 import com.ctfplatform.backend.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
@@ -9,19 +10,19 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class ChallengeHint extends BaseEntity {
+public class TryChance extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Lob
-    @Column(nullable = false)
-    private String hint;
-
     @Builder.Default
     @Column(nullable = false)
-    private Integer points = 0;
+    private Integer attempts = 0;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "team")
+    private Team team;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "challenge")
