--- src/main/java/com/ctfplatform/backend/domain/mapping/AnnouncementAuthor.java
@@ -1,28 +1,0 @@
-package com.ctfplatform.backend.domain.mapping;
-
-import com.ctfplatform.backend.domain.Announcement;
-import com.ctfplatform.backend.domain.User;
-import com.ctfplatform.backend.domain.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class AnnouncementAuthor extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "announcement")
-    private Announcement announcement;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user")
-    private User user;
-
-}
